Allow to return IPv6 address for A-record request
Add an option item only for special rules, so that
some may get IPv6 address and visit IPv6 websites normally, if:
* IPv6 on top-priority network adapter is disabled
* Other adapter(s) have IPv6 protocol enabled.

Originally, Windows will not try to request any Aaaa record
if meeting the above, which will prevent users to take advantage
of their IPv6 networks. And this PR is a work-around for it.

Note: this breaks some specification, but is indeed useful.